FINERACT-2181: Split Datatable service
