+++ fineract-core/src/main/java/org/apache/fineract/infrastructure/dataqueries/service/DatatableWriteService.java
@@ -18,25 +18,11 @@
  */
 package org.apache.fineract.infrastructure.dataqueries.service;
 
-import com.google.gson.JsonObject;
-import jakarta.validation.constraints.NotNull;
-import java.util.List;
-import java.util.Locale;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
 import org.apache.fineract.infrastructure.core.data.CommandProcessingResult;
-import org.apache.fineract.infrastructure.core.service.PagedLocalRequest;
-import org.apache.fineract.infrastructure.dataqueries.data.DatatableData;
-import org.apache.fineract.infrastructure.dataqueries.data.EntityTables;
-import org.apache.fineract.infrastructure.dataqueries.data.GenericResultsetData;
-import org.apache.fineract.portfolio.search.data.AdvancedQueryData;
-import org.springframework.data.domain.Page;
 import org.springframework.security.access.prepost.PreAuthorize;
 
-public interface ReadWriteNonCoreDataService {
-
-    List<DatatableData> retrieveDatatableNames(String appTable);
-
-    DatatableData retrieveDatatable(String datatable);
+public interface DatatableWriteService {
 
     @PreAuthorize(value = "hasAnyAuthority('ALL_FUNCTIONS', 'REGISTER_DATATABLE')")
     void registerDatatable(JsonCommand command);
@@ -50,8 +36,6 @@ public interface ReadWriteNonCoreDataService {
     @PreAuthorize(value = "hasAnyAuthority('ALL_FUNCTIONS', 'DEREGISTER_DATATABLE')")
     void deregisterDatatable(String datatable);
 
-    GenericResultsetData retrieveDataTableGenericResultSet(String datatable, Long appTableId, String order, Long id);
-
     CommandProcessingResult createDatatable(JsonCommand command);
 
     void updateDatatable(String datatableName, JsonCommand command);
@@ -62,8 +46,6 @@ public interface ReadWriteNonCoreDataService {
 
     CommandProcessingResult createNewDatatableEntry(String datatable, Long appTableId, String json);
 
-    CommandProcessingResult createPPIEntry(String datatable, Long appTableId, JsonCommand command);
-
     CommandProcessingResult updateDatatableEntryOneToOne(String datatable, Long appTableId, JsonCommand command);
 
     CommandProcessingResult updateDatatableEntryOneToMany(String datatable, Long appTableId, Long datatableId, JsonCommand command);
@@ -72,18 +54,6 @@ public interface ReadWriteNonCoreDataService {
 
     CommandProcessingResult deleteDatatableEntry(String datatable, Long appTableId, Long datatableId, JsonCommand command);
 
-    String getTableName(String Url);
+    CommandProcessingResult createPPIEntry(String datatable, Long appTableId, JsonCommand command);
 
-    String getDataTableName(String Url);
-
-    Long countDatatableEntries(String datatableName, Long appTableId, String foreignKeyColumn);
-
-    List<JsonObject> queryDataTable(@NotNull String datatable, @NotNull String columnName, String columnValue,
-            @NotNull String resultColumns);
-
-    Page<JsonObject> queryDataTableAdvanced(@NotNull String datatable, @NotNull PagedLocalRequest<AdvancedQueryData> pagedRequest);
-
-    boolean buildDataQueryEmbedded(@NotNull EntityTables entityTable, @NotNull String datatable, @NotNull AdvancedQueryData request,
-            @NotNull List<String> selectColumns, @NotNull StringBuilder select, @NotNull StringBuilder from, @NotNull StringBuilder where,
-            @NotNull List<Object> params, String mainAlias, String alias, String dateFormat, String dateTimeFormat, Locale locale);
 }
